Move ValueSources::Reportable mixin methods to AbstractValueSource
The mixin module was unnecessary and confusingly named
